fix reseeding
Using the full seed of ChaCha means that the seed doesn't change
as we run the cipher (there's just a count of how many blocks
you've evolved since the initial state).

This works with that by modifying the seed directly with XORs
against the TRNG output.

The security margin on the reseed interval is loosened up
because we've increased the security margin (and the cost) of
reseeding from the conditioned kernel TRNG-seeded CSPRNG output.