add Iox IRQ registration and event forwarding
the Iox block can register up to 8 interrupts on any of the
GPIO pins on the device. All the interrupts get OR'd together
to generate a single interrupt to the CPU, which has to be
disambiguated and forwarded on to the registering server as an
event.

Due to the protection model of Xous, the configuration of
the IRQ has to be centralized, and the handling is done with
a message. This means that the actual time from the interrupt
to the handling can be a bit variable, up to several ms, so
this method isn't suitable for hard real time.

For hard real time stuff, see the implementation in the USB
stack for Cramium, which has a very timing sensitive IRQ
handler requirement.