KB_C2 pin just died?
KB_C2 pin on this test board just got shorted to ground internally
in the chip. I think it may have been due to power glitches
when the ammeter was switching modes feeding through and causing
trouble for the system.

Anyawys, disable that row and remap some keys so we can continue
with testing. MPW boards will hopefully be retired soon enough...