move platform-specific locations into the xous-rs crates
This enables a couple of things that have been long wanting, including:

- the ability to know what PAGE_SIZE is from a single central define
- the ability to determine where certain virtual memory ranges are
located (useful for the userspace swapper).

Previously these were "hidden" from crates because they were embedded
directly in the kernel implementation module, which is not accessible
as a dependency, thus causing us to e.g. create redundant copies of
the PAGE_SIZE variable. This hopefully makes migrating to a larger
page size in some future version of the OS a bit simpler.