convert repl to use error-type returns
The main benefit of this is we don't have to remember to call
the "abort" function post-amble in case of an error: you can
just return an error type.

I took a pass at trying to convert all the argument parsing
using "idiomatic" rust. In some cases, it's nice. But in
some cases it's pretty clunky and hard to follow. I think
the clunkiness is outweighed by never having to forget to
call the abort function, and one can avoid the whole...
map_err().and_then() madness by just using more traditional
if/else with returns. I think that's a fine choice to make.