move console over to UART2 and set speed to 1Mbps
this makes on the go debugging a bit easier.

the memory initialiazation also has to be moved earlier because
the static initialization of the bool that determines which
console mode to use has to be clean early on. It may even
need to move into the asm stub, but would prefer to avoid that
if I can.